Add full expansion to reps inside sets
This means that w-lang will fully expand reps to the rep number specified in exercise.
This change should lead to the introduction of the 'fail' keyword soon.
The 'fail' keyword's purpose is to prevent the full expansion of reps, which is done
automatically otherwise.
'fail' should also be allowed inside exercise clauses as well.